Add direct customization of httpx.AsyncClient
It is now possible to provide additional configuration of
httpx.AsyncClient directly through BitcoinRpc initialization.

It is user's responsibility to conform to httpx.AsyncClient interface,
as well as the fact that authentication cannot be provided to it
directly, but through BitcoinRPC.

ImproperlyConfigured exception is raised when authentication is provided
directly to httpx.AsyncClient.

Basic tests documenting what happens if not conforming are provided.

Also, as a breaking change, all exceptions are exposed at the top level
of the package, where they should have been from the start. They are
moved to their separate ._exceptions module.