Rework BitcoinRPC initialization
Previously, `BitcoinRPC` accepted 2 arguments `host` and `port`,
creating the BTC node url itself during the initialization in
`_set_url` method.

This way turned out to be inadequate, as was shown in an example
when a library user wanted to connect to BTC node via https.

Introduced changed in this commit delegate the url creation to the
caller, via dropping the 2 arguments `host` and `port` in favour
of `url`.